refactor(esl-footnotes): move esl-note to esl-footnotes as part of component
