+++ pages/views/drafts/esl-footnotes.njk
@@ -1,7 +1,7 @@
 ---
 layout: basic
-title: ESL Notes component
-name: Esl Notes component
+title: ESL Footnotes component
+name: Esl Footnotes component
 containerCls: container
 tags: drafts
 ---
