+++ src/modules/all.less
@@ -18,6 +18,5 @@
 
 @import "./esl-forms/all.less";
 
-@import "./esl-note/core.less";
 @import "./esl-footnotes/core.less";
 @import "./esl-tooltip/core.less";
