+++ src/modules/all.ts
@@ -22,6 +22,5 @@ export * from './esl-alert/core';
 
 export * from './esl-forms/all';
 
-export * from './esl-note/core';
 export * from './esl-footnotes/core';
 export * from './esl-tooltip/core';
