+++ src/modules/esl-footnotes/core.less
@@ -1,1 +1,2 @@
 @import "./core/esl-footnotes.less";
+@import "./core/esl-note.less";
