+++ src/modules/esl-footnotes/core.ts
@@ -1,1 +1,2 @@
 export * from './core/esl-footnotes';
+export * from './core/esl-note';
