+++ src/modules/esl-footnotes/core/esl-note.less
