--- src/modules/esl-note/core.less
@@ -1,1 +1,0 @@
-@import "./core/esl-note.less";
