--- src/modules/esl-note/core.ts
@@ -1,1 +1,0 @@
-export * from './core/esl-note';
